implemented explicit null_first marker operators
The default behavior for pandas and mssql is different, thus null_first is ternary logic: True, False, don't care.
If you force null_first to something a backend is not capable of, it will throw NotImplementError. There are tests that check this behavior.
Future work:
The sign_peeler() function might be generalized better for accomodating marker operators in general since there might be coming more.