remove str_join operation
it does not seem like something very common and every backend does something
different and calls it string join / string agg or similar. The SQL backends do
a cumulative join, polars not. Maybe we can add it back in as a function only
allowed in summarise and not in mutate, since there the behaviour (apart from MSSQL,
of course) is consistent.