adapt __getattr__ on Table, remove Table __dir__
The __dir__ on a table does not have any use with complete lazy evaluation.
The available columns of a TableExpr should only be computed on demand
(e.g. when the dot is typed), otherwise, there is too much to store (we'd have
to store col names for every verb). But computing them quickly on demand is easy,
since we can just propagate the col names through the tree without passing to the
backend.