use uuids to make select checking possible
We want to eagerly check whether the user selects columns currently not in
the table.
=> We need to store a _select attr in a TableExpr.
=> We'd have to eagerly resolve the unique name of the selected columns
   to be able to do the check for columns selected on a table deeper down in
   the tree.
But this is impossible to do without significant (worst-case quadratic) overhead.
Either one has to walk up the table tree, or one needs to store a mapping from every
pair of table and column to the unique name.