make floordiv / modulo work equivalently
We do not use the python convention for floordiv (always rounding towards
negative infinity), but instead round towards zero. The SQL backends and C does it
like that. Related to that, the sign of the result of % is determined by the first
operand, not the second (again, this is SQL / C convention, in python it is determined
by the second operand). A reason for using the C convention is that

                       -(a // b) == -a // b == a // -b

which is not the case for python.