relax alias requirement for subqueries
To create a subquery, transform always needs an `alias`. Before this commit, however,
transform only considered an `alias` directly preceding the problematic verb for a
subquery. Now, also an earlier `alias` can be used to make a subquery possible. This
is nice for the following reason: If you use e.g. a `rename` or a `mutate` with only
element-wise operations after the alias, it does not affect the need for an alias. Thus
the relative placment of `alias` with resepect to such verbs should not matter.