Usability improvements (#73)
* allow Col and ColName as `rename` keys

* print more useful names for unnamed tables

* throw error on group by non-selected column

* throw error for non-aggregated cols in summarize

* add custom __repr__ for AST

This uses the ColExpr.ast_repr function for printing the inner expressions.

* add `name` and `ast_repr` verbs to documentation

* explain `pipe` parameter in docs

* add sin, cos, sqrt, etc.

* add clip

* implement new name collision resolution for join

* relax alias requirement for subqueries

To create a subquery, transform always needs an `alias`. Before this commit, however,
transform only considered an `alias` directly preceding the problematic verb for a
subquery. Now, also an earlier `alias` can be used to make a subquery possible. This
is nice for the following reason: If you use e.g. a `rename` or a `mutate` with only
element-wise operations after the alias, it does not affect the need for an alias. Thus
the relative placment of `alias` with resepect to such verbs should not matter.

* add test for join renaming

* do not print stack trace in ColExpr.__repr__

Also make all column expression __repr__`s one-liners for better readability.
Add depth parameter to ColExpr.__repr__

* improve ColExpr AST repr

* write proper Cache __repr__

* fix format string issues

* print example for failed column export

* print AST path to error source in verbs

* fix and add tests

* add self-join tests

* add Operator and ColName repr

* make verb AST repr nicer

like git log --graph

* add basic aligned functionality

* add `Series` and checks for aligned

* add tests for aligned decorator

* handle aligned in column expression export

* add tests and fix EvalAligned repr

* document @verb

* document @aligned, eval_aligned

* update documentation

* change join renamining policy

If all clashing columns are part of the join condition, only they are renamed,
regardless of how the join condition is structured.

* add cum_sum for all backends

* Fix cbrt for IBM DB2

* add pdt.rand()

* change dummy table name in DuckDbPolarsImpl

Co-authored-by: windiana42 <61181806+windiana42@users.noreply.github.com>

* add return type in polars rand()

* try RAND with seed on ibm

* Fix IBM DB2 instructions in README.md

* Fix IBM DB2 cum_sum test

* Update documentation

---------

Co-authored-by: windiana42 <windiana@users.sf.net>
Co-authored-by: windiana42 <61181806+windiana42@users.noreply.github.com>